trading-engine/requirements.txt
MetaTrader5>=5.0.45
pyyaml>=6.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
psycopg2-pool>=1.1









